requirements.txt
aiogram==3.13.1
aiosqlite
APScheduler
python-dotenv
